webpack configured, structure refactored
- webpack module bundler is configured for the entire project
- ES6 import syntax can be used easily
- webpack builds the extension into /dist folder
- folder structure refactored and organised
- use `yarn dev` for development purpose
