+++ README.md
@@ -5,6 +5,7 @@ This web extension is a JavaScript-based tool that can be installed on your brow
 ## Installation Instructions
 
 ### Mozilla Firefox 🔥
+
 1. Download the extension package.
 2. Open Mozilla Firefox and navigate to about:debugging.
 3. Click on the "This Firefox" button in the upper right corner of the screen.
@@ -13,6 +14,7 @@ This web extension is a JavaScript-based tool that can be installed on your brow
 6. You should now see the extension added to the list of installed extensions.
 
 ### Google Chrome 🌐
+
 1. Download the extension package.
 2. Open Google Chrome and navigate to chrome://extensions.
 3. Enable Developer Mode by toggling the switch in the upper right corner of the screen.
@@ -21,6 +23,7 @@ This web extension is a JavaScript-based tool that can be installed on your brow
 6. You should now see the extension added to the list of installed extensions.
 
 ### Microsoft Edge 🌊
+
 1. Download the extension package.
 2. Open Microsoft Edge and navigate to edge://extensions/.
 3. Enable Developer Mode by toggling the switch in the lower left corner of the screen.
@@ -29,10 +32,12 @@ This web extension is a JavaScript-based tool that can be installed on your brow
 6. You should now see the extension added to the list of installed extensions.
 
 ## Usage Instructions 📖
+
 Once the extension is installed, you can start using it immediately. Simply click on the extension icon in your browser toolbar to access its features. The specific
- functionality of the extension will depend on its purpose, so please refer to its documentation for more information.
+functionality of the extension will depend on its purpose, so please refer to its documentation for more information.
 
 ## Contributing 🤝
+
 Contributions are always welcome! If you would like to contribute to the development of this web extension, please follow these steps:
 
 1. Fork the repository to your own GitHub account.
@@ -42,6 +47,15 @@ Contributions are always welcome! If you would like to contribute to the develop
 5. Submit a pull request to the original repository and explain the changes you made.
 
 I would recommend `web-ext` for development in Firefox, it significantly decrease development times. Devs can install it using `npm install --global web-ext`.
+
+## Development
+
+1. Make sure you use `yarn` and not `npm` while installing the packages
+2. Webpack is configured for the project
+3. Do not edit the manifest.json, webpack.config.js files
+4. Clone the repo
+5. Run `yarn install` in the terminal to install the required dependencies
+6. Run `yarn dev` to test the extension. A new firefox window will open with the extension loaded.
 
 ## License:
 
@@ -55,17 +69,16 @@ Under this license, you are free to:
 Under the following terms:
 
 - Attribution: You must give appropriate credit, provide a link to the license, and indicate if changes were made. You may do so in any reasonable manner, but not in any way
- that suggests the licensor endorses you or your use.
+  that suggests the licensor endorses you or your use.
 - NonCommercial: You may not use the material for commercial purposes.
 - ShareAlike: If you remix, transform, or build upon the material, you must distribute your contributions under the same license as the original.
 
 For more details, please refer to the full license text at https://creativecommons.org/licenses/by-nc-sa/4.0/legalcode.
 
-
 ## Resources 🔍
+
 - [Google Chrome Developer Documentation](https://developer.chrome.com/docs/extensions/)
 - [Mozilla Firefox Extension Workshop](https://extensionworkshop.com/)
 - [Microsoft Edge Extension Documentation](https://docs.microsoft.com/en-us/microsoft-edge/extensions-chromium/)
 - [Mozilla Extension Documentation](https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions)
 - [Microsoft Edge Getting Started](https://learn.microsoft.com/en-us/microsoft-edge/extensions-chromium/)
-
