+++ package.json
@@ -1,0 +1,13 @@
+{
+  "scripts": {
+    "build": "webpack",
+    "start:firefox": "cd dist && web-ext run",
+    "dev": "yarn build && yarn start:firefox"
+  },
+  "devDependencies": {
+    "copy-webpack-plugin": "^11.0.0",
+    "watch": "^1.0.2",
+    "webpack": "^5.88.2",
+    "webpack-cli": "^5.1.4"
+  }
+}
