+++ src/filler/engines/fields-extractor-engine.js
@@ -1,0 +1,7 @@
+export class FieldsExtractorEngine {
+  // Extracts the questions and description from the DOM object and returns it
+  // It might also extract the options in case of MCQs or other types, where answers do
+  // play a  critical role
+
+  constructor() {}
+}
