+++ src/popup/popup.html
@@ -1,0 +1,25 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Document</title>
+  </head>
+  <body>
+    <div class="root">
+      <div style="background-color: aquamarine; height: 200px; width: 200px">
+        <center
+          style="
+            padding-top: 80px;
+            color: grey;
+            font-weight: 600;
+            font-family: 'Lucida Sans', 'Lucida Sans Regular', 'Lucida Grande',
+              'Lucida Sans Unicode', Geneva, Verdana, sans-serif;
+          "
+        >
+          This is test Extension Page yayy
+        </center>
+      </div>
+    </div>
+  </body>
+</html>
